+++ ui/apps/platform/src/Containers/Collections/RuleSelector/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './RuleSelector';
+export type { RuleSelectorProps } from './RuleSelector';
